delete categories

+++ app/api/[[...route]]/categories.ts
@@ -4,6 +4,8 @@ import { Hono } from "hono";
 import { db } from "@/db/drizzle";
 import { categories } from "@/db/schema";
 import { count, eq, inArray } from "drizzle-orm";
+import { verifyAuth } from "@hono/auth-js";
+import { deleteFromS3 } from "@/lib/s3-upload";
 
 const app = new Hono()
   .get(
@@ -73,7 +75,8 @@ const app = new Hono()
     }
   )
   .post(
-    "/delete-categories",
+    "/delete",
+    verifyAuth(),
     zValidator(
       "json",
       z.object({
@@ -81,14 +84,47 @@ const app = new Hono()
       })
     ),
     async (c) => {
+      const auth = c.get("authUser");
       const values = c.req.valid("json");
+
+      if (!auth || !auth.token?.sub) {
+        return c.json({ error: "Unauthorized" }, 401);
+      }
+
+      const uploadedFiles: string[] = [];
+
+      const existingImages = await db
+        .select({
+          image: categories.image,
+        })
+        .from(categories)
+        .where(inArray(categories.id, values.ids));
+
+      if (!existingImages || existingImages.length === 0) {
+        return c.json({ error: "Failed to fetch images" }, 500);
+      }
+
+      for (const image of existingImages) {
+        if (image.image) {
+          const imageName = image.image.split("/").pop() || "";
+          uploadedFiles.push(imageName);
+        }
+      }
+
+      for (const fileName of uploadedFiles) {
+        const { success, message } = await deleteFromS3(fileName);
+        if (!success) {
+          console.error(`Falha ao deletar ${fileName}: ${message}`);
+          return c.json({ error: message }, 500);
+        }
+      }
 
       const data = await db
         .delete(categories)
         .where(inArray(categories.id, values.ids));
 
       if (!data) {
-        return c.json({ error: "Failed to delete categories" }, 500);
+        return c.json({ error: "Falha ao deletar as categorias" }, 500);
       }
 
       return c.json({ data });
@@ -109,10 +145,31 @@ const app = new Hono()
         return c.json({ error: "Missing id" }, 400);
       }
 
+      const [existingImage] = await db
+        .select({
+          image: categories.image,
+        })
+        .from(categories)
+        .where(eq(categories.id, id));
+
+      if (!existingImage || !existingImage.image) {
+        console.error("Failed to fetch image");
+        return c.json({ error: "Falha ao deletar a categoria" }, 500);
+      }
+
+      const fileName = existingImage.image.split("/").pop() || "";
+
+      const { success, message } = await deleteFromS3(fileName);
+      if (!success) {
+        console.error(`Falha ao deletar ${fileName}: ${message}`);
+        return c.json({ error: message }, 500);
+      }
+
       const data = await db.delete(categories).where(eq(categories.id, id));
 
       if (!data) {
-        return c.json({ error: "Failed to delete category" }, 500);
+        console.error("Failed to delete category");
+        return c.json({ error: "Falha ao deletar a categoria" }, 500);
       }
 
       return c.json({ data });
